app/routing/semantic_classifier.py
"""
Zero-Cost Local Semantic Classification
Uses embedding similarity instead of Gemini API calls for task categorization.
Reduces routing latency: ~1000ms → ~5ms | Cost: ~$0.004 → $0.00
"""

import numpy as np
from typing import Dict, Tuple
from app.embedding_engine import generate_vector

# Define category embeddings - these are semantic anchors for each task type
CATEGORY_DEFINITIONS = {
    "CODE": "write program code software development debugging algorithms functions classes",
    "ANALYSIS": "analyze data statistics research insights patterns trends breakdown explain",
    "CHAT": "conversation chat discussion casual dialogue talk interact communicate",
    "CREATIVE": "write poetry story fiction creative writing novel dialogue screenplay art",
    "EXTRACTION": "extract information data parsing get retrieve convert format parsing",
    "UTILITY": "convert format calculate compute calculate helper tool utility calculator",
    "AGENTS": "agent autonomous tool use workflow planning reasoning agentic behavior",
    "IMAGE": "image generation visual create picture draw design graphic art visual creation",
    "VIDEO": "video generation create movie animation footage film create video motion",
    "AUDIO": "audio speech sound music transcribe voice generation spoken audio",
    "MULTIMODAL": "multimodal cross-modal combine image text video audio multiple modalities"
}

class SemanticClassifier:
    """
    Local semantic classification using embedding similarity.
    Zero API calls, ~5ms latency, $0.00 cost.
    """
    
    def __init__(self):
        """Initialize category embeddings."""
        self.category_embeddings: Dict[str, np.ndarray] = {}
        self._build_category_embeddings()
    
    def _build_category_embeddings(self):
        """Generate embeddings for each category definition."""
        print("[*] Initializing semantic classifier...")
        for category, definition in CATEGORY_DEFINITIONS.items():
            embedding = generate_vector(definition)
            if embedding:
                self.category_embeddings[category] = np.array(embedding)
                print(f"    [OK] {category:12} embedding loaded (768-dim)")
        print(f"[DONE] Semantic classifier ready ({len(self.category_embeddings)} categories)\n")
    
    def classify_prompt(self, prompt: str) -> Tuple[str, float]:
        """
        Classify prompt into a category using cosine similarity.
        
        Args:
            prompt: User's input prompt
        
        Returns:
            (category: str, confidence: float 0-1)
        
        Performance:
            - Latency: ~5ms (local only)
            - Cost: $0.00 (no API calls)
        """
        # Generate embedding for the prompt
        prompt_embedding = generate_vector(prompt)
        if not prompt_embedding:
            return "UTILITY", 0.5  # Safe fallback
        
        prompt_vec = np.array(prompt_embedding)
        
        # Compute cosine similarity to all category embeddings
        scores = {}
        for category, category_vec in self.category_embeddings.items():
            # Cosine similarity: (A·B) / (||A|| * ||B||)
            similarity = np.dot(prompt_vec, category_vec) / (
                np.linalg.norm(prompt_vec) * np.linalg.norm(category_vec)
            )
            scores[category] = similarity
        
        # Get best match
        best_category = max(scores, key=scores.get)
        best_score = scores[best_category]
        
        # Normalize score to 0-1 range (similarity is in [-1, 1], shift to [0, 1])
        normalized_score = (best_score + 1.0) / 2.0
        
        return best_category, normalized_score
    
    def classify_with_top_k(self, prompt: str, k: int = 5) -> Dict[str, float]:
        """
        Classify prompt and return top-K category scores.
        
        Args:
            prompt: User's input prompt
            k: Number of top categories to return
        
        Returns:
            Dict mapping category → confidence score, sorted descending
        """
        prompt_embedding = generate_vector(prompt)
        if not prompt_embedding:
            return {"UTILITY": 0.5}
        
        prompt_vec = np.array(prompt_embedding)
        
        # Compute similarity scores
        scores = {}
        for category, category_vec in self.category_embeddings.items():
            similarity = np.dot(prompt_vec, category_vec) / (
                np.linalg.norm(prompt_vec) * np.linalg.norm(category_vec)
            )
            # Normalize to [0, 1]
            normalized = (similarity + 1.0) / 2.0
            scores[category] = normalized
        
        # Sort by score and return top-K
        sorted_scores = sorted(scores.items(), key=lambda x: x[1], reverse=True)
        return {cat: score for cat, score in sorted_scores[:k]}
    
    def classify_with_complexity(self, prompt: str) -> Tuple[str, float, str]:
        """
        Classify prompt into category AND estimate complexity level.
        
        Args:
            prompt: User's input prompt
        
        Returns:
            (category, confidence, complexity_label)
            where complexity_label is "EASY", "MEDIUM", or "HARD"
        """
        category, confidence = self.classify_prompt(prompt)
        
        # Estimate complexity based on prompt length, keywords
        complexity_score = self._estimate_complexity(prompt)
        
        if complexity_score < 4.0:
            complexity_label = "EASY"
        elif complexity_score < 7.0:
            complexity_label = "MEDIUM"
        else:
            complexity_label = "HARD"
        
        return category, confidence, complexity_label
    
    def _estimate_complexity(self, prompt: str) -> float:
        """
        Estimate task complexity (1.0 to 10.0) based on prompt characteristics.
        
        Heuristics:
        - Length: longer = more complex
        - Keywords: technical terms = higher complexity
        - Punctuation: more punctuation = more specific = complex
        """
        score = 1.0
        
        # Length factor (max +3 points)
        word_count = len(prompt.split())
        score += min(word_count / 100, 3.0)
        
        # Technical keywords (+0.5 each, max +3 points)
        technical_keywords = [
            "algorithm", "optimization", "architecture", "regex", "async", 
            "concurrency", "quantum", "machine learning", "neural", "transformer",
            "distributed", "microservice", "containerization", "kubernetes",
            "api gateway", "load balancer", "cache", "database"
        ]
        for keyword in technical_keywords:
            if keyword.lower() in prompt.lower():
                score += 0.5
        
        # Code complexity indicators
        if any(indicator in prompt.lower() for indicator in ["debug", "fix", "error"]):
            score += 1.0
        
        if any(indicator in prompt.lower() for indicator in ["design", "architecture", "pattern"]):
            score += 1.5
        
        # Normalize to 1.0-10.0
        return min(max(score, 1.0), 10.0)


# Global instance
_classifier = None

def get_semantic_classifier() -> SemanticClassifier:
    """Get or initialize the global semantic classifier."""
    global _classifier
    if _classifier is None:
        _classifier = SemanticClassifier()
    return _classifier
